+++ extensions/ql-vscode/test/unit-tests/common/location.test.ts
@@ -1,42 +1,29 @@
+import { tryGetRemoteLocation } from "../../../src/common/bqrs-utils";
 import {
-  tryGetRemoteLocation,
-  tryGetResolvableLocation,
-} from "../../../src/common/bqrs-utils";
+  UrlValue,
+  UrlValueResolvable,
+} from "../../../src/common/raw-result-types";
 
-describe("processing string locations", () => {
-  it("should detect Windows whole-file locations", () => {
-    const loc = "file://C:/path/to/file.ext:0:0:0:0";
-    const wholeFileLoc = tryGetResolvableLocation(loc);
-    expect(wholeFileLoc).toEqual({ uri: "C:/path/to/file.ext" });
-  });
-  it("should detect Unix whole-file locations", () => {
-    const loc = "file:///path/to/file.ext:0:0:0:0";
-    const wholeFileLoc = tryGetResolvableLocation(loc);
-    expect(wholeFileLoc).toEqual({ uri: "/path/to/file.ext" });
-  });
-
-  it("should detect Unix 5-part locations", () => {
-    const loc = "file:///path/to/file.ext:1:2:3:4";
-    const wholeFileLoc = tryGetResolvableLocation(loc);
-    expect(wholeFileLoc).toEqual({
-      uri: "/path/to/file.ext",
-      startLine: 1,
-      startColumn: 2,
-      endLine: 3,
-      endColumn: 4,
-    });
-  });
-  it("should ignore other string locations", () => {
-    for (const loc of ["file:///path/to/file.ext", "I am not a location"]) {
-      const wholeFileLoc = tryGetResolvableLocation(loc);
-      expect(wholeFileLoc).toBeUndefined();
-    }
-  });
-});
-
-describe("getting links to remote (GitHub) locations", () => {
+describe("tryGetRemoteLocation", () => {
   it("should return undefined if resolvableLocation is undefined", () => {
-    const loc = "not a location";
+    const loc = undefined;
+    const fileLinkPrefix = "";
+    const sourceLocationPrefix = "";
+
+    const link = tryGetRemoteLocation(
+      loc,
+      fileLinkPrefix,
+      sourceLocationPrefix,
+    );
+
+    expect(link).toBeUndefined();
+  });
+
+  it("should return undefined if resolvableLocation is string", () => {
+    const loc: UrlValue = {
+      type: "string",
+      value: "not a location",
+    };
     const fileLinkPrefix = "";
     const sourceLocationPrefix = "";
 
@@ -50,7 +37,8 @@ describe("getting links to remote (GitHub) locations", () => {
   });
 
   it("should return undefined if resolvableLocation has the wrong format", () => {
-    const loc = {
+    const loc: UrlValueResolvable = {
+      type: "lineColumnLocation",
       uri: "file:/path/to/file.ext",
       startLine: 194,
       startColumn: 18,
@@ -70,7 +58,8 @@ describe("getting links to remote (GitHub) locations", () => {
   });
 
   it("should return a remote file ref if the sourceLocationPrefix and resolvableLocation match up", () => {
-    const loc = {
+    const loc: UrlValueResolvable = {
+      type: "lineColumnLocation",
       uri: "file:/home/foo/bar/path/to/file.ext",
       startLine: 194,
       startColumn: 18,
@@ -92,7 +81,8 @@ describe("getting links to remote (GitHub) locations", () => {
   });
 
   it("should return undefined if the sourceLocationPrefix is missing and resolvableLocation doesn't match the default format", () => {
-    const loc = {
+    const loc: UrlValueResolvable = {
+      type: "lineColumnLocation",
       uri: "file:/home/foo/bar/path/to/file.ext",
       startLine: 194,
       startColumn: 18,
@@ -112,7 +102,8 @@ describe("getting links to remote (GitHub) locations", () => {
   });
 
   it("should return a remote file ref if the sourceLocationPrefix is missing, but the resolvableLocation matches the default format", () => {
-    const loc = {
+    const loc: UrlValueResolvable = {
+      type: "lineColumnLocation",
       uri: "file:/home/runner/work/foo/bar/path/to/file.ext",
       startLine: 194,
       startColumn: 18,